requirements.txt
langchain_core
langchain
pydantic
python-dotenv
streamlit
langchain_community
langserve
sse_starlette
pymupdf
faiss-cpu
groq
beautifulsoup4
langchain-groq
langchainhub
sentence_transformers